// challenge-zafirus/src/app/interfaces/character.interface.ts
//Creo todos los modelos ya que trae esta cantidad de datos la respuesta.
export interface CharactersResponse {
    code:             number;
    status?:          string;
    copyright?:       string;
    attributionText?: string;
    attributionHTML?: string;
    etag?:            string;
    data:             DataCharactersResponse;
}

export interface DataCharactersResponse {
    offset:  number;
    limit:   number;
    total:   number;
    count:   number;
    results: Result[];
}

export interface Result {
    id:          number;
    name:        string;
    description: string;
    modified:    string;
    thumbnail:   Thumbnail;
    resourceURI: string;
    comics:      Comics;
    series:      Comics;
    stories:     Stories;
    events:      Comics;
    urls:        URL[];
}

export interface Comics {
    available:     number;
    collectionURI: string;
    items:         ComicsItem[];
    returned:      number;
}

export interface ComicsItem {
    resourceURI: string;
    name:        string;
}

export interface Stories {
    available:     number;
    collectionURI: string;
    items:         StoriesItem[];
    returned:      number;
}

export interface StoriesItem {
    resourceURI: string;
    name:        string;
    type:        string;
}

export interface Thumbnail {
    path:      string;
    extension: string;
}

export interface URL {
    type: string;
    url:  string;
}